+++ src/common/cleanIpcErrorMessage.ts
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) 2022 Nordic Semiconductor ASA
+ * Copyright (c) 2025 Nordic Semiconductor ASA
  *
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-export const cleanIpcErrorMessage = (message: string, replacecment = '') =>
+export default (message: string, replacecment = '') =>
     message.replace(
         /Error invoking remote method '.*?': *(Error:)? */,
         replacecment
